pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "block-smc"
version = "0.1.0"
description = "Block SMC with learned twist functions and EM boundary learning"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "numpy",
    "genlm-control",
    "llamppl>=0.2.2",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
]
dev = [
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
